Hub card compiler tests passing
