+++ packages/compiler-server/src/server.ts
@@ -6,11 +6,10 @@ import sane from 'sane';
 import { setupWatchers } from './watcher';
 import { CardStackContext, HubServerConfig } from '@cardstack/hub/interfaces';
 import { setupCardBuilding } from './context/card-building';
-import { cardRoutes } from '../../hub/routes/card-routes';
 
 export class HubServer {
   static async create(options: Partial<HubServerConfig>): Promise<HubServer> {
-    let { realms, cardCacheDir, routeCard } = options;
+    let { realms, cardCacheDir } = options;
 
     // NOTE: PR defining how to do types on Koa context and application:
     // https://github.com/DefinitelyTyped/DefinitelyTyped/pull/31704
@@ -21,11 +20,7 @@ export class HubServer {
 
     setupCardBuilding(app, { realms: realms!, cardCacheDir: cardCacheDir! });
 
-    let router = await cardRoutes(app.context, routeCard);
-    app.use(router.routes());
-    app.use(router.allowedMethods());
-
-    return new this(app, options);
+    return new this(app, options as HubServerConfig);
   }
 
   private watchers: sane.Watcher[] | undefined;
