+++ packages/core/src/errors.ts
@@ -1,0 +1,4 @@
+class CompilerError extends Error {
+  isCompilerError = true;
+}
+export class InvalidKeysError extends CompilerError {}
