+++ packages/hub/middleware/error.ts
@@ -4,17 +4,27 @@ export default async function error(ctxt: Koa.Context, next: Koa.Next) {
   try {
     return await next();
   } catch (err: any) {
-    if (!err.isCardstackError) {
+    if (!err.isCardstackError && !err.isCompilerError) {
       throw err;
     }
     if (err.status === 500) {
       console.error(`Unexpected error: ${err.status} - ${err.message}\n${err.stack}`); // eslint-disable-line no-console
     }
+    if (err.isCompilerError) {
+      // switch (instanceof err) {
+      //   case 'InvalidKeysError':
+      err.status = 400;
+      err.title = 'Bad Request';
+      //   break;
+
+      // default:
+      //   break;
+    }
     let errors = [err];
     if (err.additionalErrors) {
       errors = errors.concat(err.additionalErrors);
     }
+    ctxt.status = err.status;
     ctxt.body = { errors };
-    ctxt.status = errors[0].status;
   }
 }
