+++ packages/hub/node-tests/helpers/cards.ts
@@ -2,10 +2,9 @@ import { Project } from 'scenario-tester';
 import Mocha from 'mocha';
 import tmp from 'tmp';
 import { join } from 'path';
-import { ensureDirSync, outputJSONSync } from 'fs-extra';
 
 import { RawCard, RealmConfig } from '@cardstack/core/src/interfaces';
-import { wireItUp } from '../../main';
+import { HubServer, wireItUp } from '../../main';
 import { CardCacheConfig } from '../../services/card-cache-config';
 import CardBuilder from '../../services/card-builder';
 import RealmManager from '../../services/realm-manager';
@@ -20,8 +19,13 @@ export const TEST_REALM = 'https://cardstack.local';
 
 class TestCardCacheConfig extends CardCacheConfig {
   tmp = tmp.dirSync();
+
   get root() {
     return this.tmp.name;
+  }
+
+  get cacheDirectory() {
+    return join(this.root, 'node_modules', this.packageName);
   }
   cleanup() {
     this.tmp.removeCallback();
@@ -61,13 +65,13 @@ export class ProjectTestRealm extends FSRealm {
   }
 }
 
+function resolveCard(root: string, modulePath: string): string {
+  return require.resolve(modulePath, { paths: [root] });
+}
+
 export function setupCardBuilding(mochaContext: Mocha.Suite) {
   let realms: RealmManager, builder: CardBuilder;
   let cardCache: CardCache, cardCacheConfig: TestCardCacheConfig;
-
-  function resolveCard(modulePath: string): string {
-    return require.resolve(modulePath, { paths: [cardCacheConfig.root] });
-  }
 
   function createRealm(name: string): ProjectTestRealm {
     return realms.createRealm({ url: name }, ProjectTestRealm);
@@ -77,9 +81,6 @@ export function setupCardBuilding(mochaContext: Mocha.Suite) {
     let container = wireItUp((registry: Registry) => {
       registry.register('card-cache-config', TestCardCacheConfig);
     });
-
-    let cardCacheConfig = await container.lookup('card-cache-config');
-    ensureDirSync(cardCacheConfig.cacheDirectory);
 
     cardCache = await container.lookup('card-cache');
     realms = await container.lookup('realm-manager');
@@ -102,25 +103,49 @@ export function setupCardBuilding(mochaContext: Mocha.Suite) {
     getRealmManager() {
       return realms;
     },
-    resolveCard,
+    resolveCard(module: string) {
+      return resolveCard(cardCacheConfig.root, module);
+    },
     createRealm,
   };
 }
 
-export const MINIMAL_PACKAGE = {
-  name: '@cardstack/compiled',
-  exports: {
-    '.': {
-      browser: './browser',
-      default: './node',
-    },
-    './*': {
-      browser: './browser/*',
-      default: './node/*',
-    },
-  },
-};
+export function setupCardServer(mochaContext: Mocha.Suite) {
+  let server: HubServer, cardCache: CardCache, cardCacheConfig: TestCardCacheConfig;
 
-export function createMinimalPackageJSON(cardCacheDir: string): void {
-  outputJSONSync(join(cardCacheDir, 'package.json'), MINIMAL_PACKAGE);
+  function createRealm(name: string): ProjectTestRealm {
+    return server.builder.realmManager.createRealm({ url: name }, ProjectTestRealm);
+  }
+
+  mochaContext.beforeEach(async function () {
+    server = await HubServer.create({
+      registryCallback: (registry: Registry) => {
+        registry.register('card-cache-config', TestCardCacheConfig);
+      },
+    });
+
+    cardCache = await server.container.lookup('card-cache');
+    cardCacheConfig = (await server.container.lookup('card-cache-config')) as TestCardCacheConfig;
+  });
+
+  mochaContext.afterEach(function () {
+    server.teardown();
+    cardCache.cleanup();
+  });
+
+  return {
+    createRealm,
+    getCardCache() {
+      return cardCache;
+    },
+    getServer() {
+      return server;
+    },
+    resolveCard(module: string) {
+      return resolveCard(cardCacheConfig.root, module);
+    },
+    async setRoutingCard(cardURL: string) {
+      return (await server.container.lookup('card-routes')).setRoutingCard(cardURL);
+    },
+  };
 }
