+++ packages/hub/utils/serialization.ts
